move exportable stuff into package dir (as discussed with @martingartonft)
